MetaDataExtractionSystem/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pytesseract
Pillow
openai
pydantic
httpx
python-dotenv
path